Fix(build) : FlexSearch.Server wasn't loading in VS
Error desc: The project 'FlexSearch.Server.fsproj' could not be opened
because opening it would cause a folder to be rendered multiple times in
the solution explorer. One such problematic item is
'Web\Reports\DuplicateDetectionTemplate.html'. To open this project in
Visual Studio, first edit the project file and fix the problem.